Improve Spanish translation (#65)
Hello, congratulations for this awesome project.
I am a beginner and wanted to contribute, also in the frame of Hacktoberfest. Being a native Spanish speaker, I can offer to review the Spanish translation of the README and make a some improvements so the language flows more naturally and is more pleasant to read. Although it was easy to understand the original version there were a number of typos and expressions that didnt quite sound right.